Implement spliting the rendering in multiple draw calls
Before, if the number of textures used was greater than
MAX_TEXTURE_IMAGE_UNITS, the program would panic.

And if the number of sprites were greater than 2^16/4, the indices in
the indice buffer would overflow, not rendering the remainer of the
sprites.

That was fixed by spliting the rendering when any of the above
conditions is met.